Readme madness
With the repo being made the official selfhosting repo for mee6, I jumped at the opportunity to improve this repo. I added sudo in some places to README, and opened up the possibility of using alternative programs in the script (different download agent, different web server, etc.). I added the export variables to their own file. so that one can easily amend the variables to .bashrc.